refactor: extract jsonl processor from cmd
