+++ examples/result/small.schema
@@ -1,1 +1,1 @@
-Record({"age": Atomic(Int()), "name": Atomic(Str()), "surname": Atomic(Str())})
+Unknown()
