+++ python/jskiner/cmd.py
@@ -1,10 +1,6 @@
 import argparse
 import subprocess
-import os
-import shutil
-from .jskiner import InferenceEngine
-
-exec("from .schema import *")
+from .jsonl import JsonlProcessor
 
 
 def get_args():
@@ -67,71 +63,8 @@ def run() -> None:
     args = get_args()
     if args.verbose:
         print(f"Loading {args.jsonl}")
-    if args.split == 1:
-        schema_str = get_schema_from_jsonl(args.jsonl, worker_cnt=args.nworkers)
-    else:
-        schema_str = get_schema_batchwise(
-            args.jsonl, args.split_path, args.split, verbose=args.verbose
-        )
+    schema_str = JsonlProcessor(args).run()
     store(schema_str, output_path=args.out, verbose=args.verbose, format=args.format)
-
-
-def get_schema_batchwise(src_path, split_path, split_cnt, verbose=False):
-    try:
-        refresh_split_path(split_path)
-        split(src_path, split_path, split_cnt)
-        schema = eval("Unknown()")
-        file_iter = os.listdir(split_path)
-        if verbose:
-            try:
-                import tqdm
-            except ImportError:
-                subprocess.run(["pip", "install", "tqdm"])
-            file_iter = tqdm.tqdm(file_iter)
-        for file_name in file_iter:
-            selected_path = f"{split_path}/{file_name}"
-            if verbose:
-                print("Start Inferencing", selected_path)
-            schema_str = get_schema_from_jsonl(selected_path)
-            schema |= eval(schema_str)
-            if verbose:
-                print("Finish Inferencing", selected_path)
-        schema_str = schema.__repr__()
-        return schema_str
-    except BaseException as e:
-        with open("log", "w") as f:
-            f.write(schema_str)
-        raise e
-    finally:
-        refresh_split_path(split_path)
-
-
-def refresh_split_path(path):
-    if not os.path.exists(path):
-        os.mkdir(path)
-    else:
-        shutil.rmtree(path)
-        os.mkdir(path)
-
-
-def split(src_path, split_path, split_cnt):
-    total = get_total_json_count(src_path)
-    cnt_per_file = int(total / split_cnt)
-    subprocess.run(["split", "-l", str(cnt_per_file), src_path, split_path + "/"])
-
-
-def get_total_json_count(path):
-    out = subprocess.check_output(["wc", "-l", path])
-    total = int(out.decode("utf-8").split(path)[0])
-    return total
-
-
-def get_schema_from_jsonl(jsonl_path, worker_cnt=1):
-    with open(jsonl_path, "r") as f:
-        json_list = [x for x in f]
-    engine = InferenceEngine(worker_cnt)
-    schema_str = engine.run(json_list)
-    return schema_str
 
 
 def store(schema_str, output_path="out.schema", verbose=False, format=True):
@@ -142,7 +75,7 @@ def store(schema_str, output_path="out.schema", verbose=False, format=True):
             print("Result saved into", output_path)
         if format:
             try:
-                exec('import black')
+                exec("import black")
             except ImportError:
                 subprocess.run(["pip", "install", "black"])
             subprocess.run(["black", output_path])
